now it's a lot more accurate. introduce Impossible, which indicates:
the combination just doesn't exist! it happens when there's an
impossible substraction, where min > max. previously, i just mark it an
EmptyClue, but EmptyClue counts 1, not 0. now if we're introducing
Impossible, which counts 0, and the entire multiplication would be 0!
that way, we can remove impossible combinations.

also, move count to Clue instead of ExclusiveClue. this... is only
to ease the type check. we're not using ConjunctedClue#count and
SubtractedClue#count anyway. if that happens, it should be a bug!